move: partner/interfaces -> common/models
