+++ api/common/models/product.ts
