+++ api/partner/interfaces/index.ts
@@ -1,9 +1,9 @@
 import * as auth from "./auth";
 import * as error from "./error";
-import * as event from "./event";
-import * as order from "./order";
-import * as orderProduct from "./order-product";
-import * as product from "./product";
+import * as event from "../../common/models/event";
+import * as order from "../../common/models/order";
+import * as orderProduct from "../../common/models/order-product";
+import * as product from "../../common/models/product";
 import * as user from "./user";
 
 export { auth };
